requirements.txt
jupyter==1.0.0
python-dotenv==1.0.0
openai==1.8.0
chromadb==0.4.22
pandas==2.0.3
pyarrow==15.0.0
streamlit==1.30.0